refactor: split transformer node dispatch
